Clean up menu interfaces
