+++ src/ui/tsconfig.json
@@ -3,7 +3,6 @@
     "declaration": true,
     "noImplicitAny": true,
     "noEmitOnError": true,
-    "sourceMap": true,
     "module": "commonjs",
     "moduleResolution": "node",
     "target": "ES5",
